requirements.txt
matplotlib
tqdm
lightning
torchvision
pytorch-fid